Polices para gerentes

+++ app/Http/Controllers/DeletarProdutoController.php
@@ -8,11 +8,15 @@ use App\Models\Produto;
 class DeletarProdutoController extends Controller
 {
     public function deletar($id){
+        $this->authorize("cadastrar", \App\Models\Funcionario::class);
+
         $produto = produto::findOrFail($id);
         return view('/deletarProduto', ['produto' => $produto]);
     }
 
     public function excluir($id) {
+        $this->authorize("cadastrar", \App\Models\Funcionario::class);
+
         $produto = produto::findOrFail($id);
         $produto->delete();
